requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.2
torchvision==0.16.2

# YOLO & Computer Vision
ultralytics==8.4.14
lap==0.5.12
timm==0.9.16
# bytetrack pin removed — install manually if needed (incompatible wheel on macOS/Python 3.11)
# bytetrack==1.0.5

# GUI
PySide6==6.6.1

# Database & API
supabase==2.28.0
# Let supabase manage compatible versions of its core sub-dependencies.
# postgrest-py
# python-gotrue
# realtime-py

# Configuration
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.0
Pillow==10.0.1
scikit-learn==1.3.2
scipy==1.11.4

# Logging & Monitoring
python-logging-loki==0.3.1  # optional: Grafana Loki

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
